+++ server/datas/mapping.data.json
@@ -1,0 +1,14 @@
+{
+    "hotels": [
+        
+    ],
+    "users": [
+
+    ],
+    "visits": [
+
+    ],
+    "vehicules": [
+
+    ]
+}
