+++ server/model/urgence.model.js
@@ -1,5 +1,6 @@
 const mongoose = require('mongoose');
 const Hotel = require("./hotel.model");
+const Equipe = require("./assoc_user_user.model");
 
 const Schema = mongoose.Schema;
 
@@ -7,6 +8,10 @@ const urgenceSchema = new Schema({
     hotel_id : {
         type: Schema.Types.ObjectId, 
         ref: 'Hotel', 
+    },
+    equipe_id : {
+        type: Schema.Types.ObjectId,
+        ref: 'Equipe'
     },
     resume: {
         type: String, 
@@ -19,7 +24,7 @@ const urgenceSchema = new Schema({
         required: true,
         trim: true,
         maxlength: 3600
-    }
+    },
 })
 
 const Urgence = mongoose.model('Urgence', urgenceSchema)
