+++ server/routes/feature.gestion_urgence/crudUrgence.routes.js
@@ -40,7 +40,8 @@ router.route('/get/:id').get((req, res) => {
 router.route('/add').post((req, res) => {
     //creer model Hotel
     const urgence = new Urgence({
-        hotel_id : req.body.id_hotel, 
+        hotel_id : req.body.id_hotel,
+        equipe_id : req.body.id_equipe,  
         resume : req.body.resume, 
         detail : req.body.detail 
     })
